Editing part D services

+++ src/main/java/D/Service/CourseService.java
@@ -5,9 +5,9 @@ import D.Entities.Course;
 import java.util.List;
 
 public interface CourseService {
-    Long create (Course course);
-    Course findById(Long id);
+    Integer create (Course course);
+    Course findById(Integer id);
     List<Course> findAll();
     void Update(Course course);
-    void Delete(Long id);
+    void Delete(Integer id);
 }
