+++ src/main/java/D/Service/Impl/EmployeeServiceImpl.java
@@ -1,6 +1,7 @@
 package D.Service.Impl;
 
 import D.Entities.Employee;
+import D.Entities.Prof;
 import D.Repository.Impl.EmployeeRepositoryImpl;
 import D.Service.EmployeeService;
 
@@ -15,24 +16,24 @@ public class EmployeeServiceImpl implements EmployeeService {
         employeeRepository=new EmployeeRepositoryImpl();
     }
 
+
     @Override
-    public Long login(Employee employee) {
-      List<Employee> employees=  employeeRepository.findAll();
-List<Employee> employees1=      employees.stream().filter(x-> Objects.equals(x.getUsername(), employee.getUsername()) && Objects.equals(x.getPassword(), employee.getPassword())).collect(Collectors.toList());
-    if(employees1.size()>0){
-        return employees1.get(0).getId();
-    }
-      return 0l;
+    public Integer login(Employee employee) {
+        List<Employee> employees =  employeeRepository.findAll();
+        List<Employee> employees1=      employees.stream().filter(x-> Objects.equals(x.getUsername(), employee.getUsername()) && Objects.equals(x.getPassword(), employee.getPassword())).collect(Collectors.toList());
+        if(employees1.size()>0){
+            return employees1.get(0).getId();
+        }
+        return 0;
     }
 
     @Override
-    public Long create(Employee employee) {
-
+    public Integer create(Employee employee) {
         return employeeRepository.create(employee);
     }
 
     @Override
-    public Employee findById(Long id) {
+    public Employee findById(Integer id) {
         return employeeRepository.findById(id);
     }
 
@@ -43,11 +44,11 @@ List<Employee> employees1=      employees.stream().filter(x-> Objects.equals(x.g
 
     @Override
     public void Update(Employee employee) {
-    employeeRepository.Update(employee);
+employeeRepository.Update(employee);
     }
 
     @Override
-    public void Delete(Long id) {
-       employeeRepository.Delete(id);
+    public void Delete(Integer id) {
+employeeRepository.Delete(id);
     }
 }
