+++ src/main/java/D/Service/ProfService.java
@@ -2,5 +2,5 @@ package D.Service;
 
 import D.Entities.Prof;
 
-public interface ProfService extends BaseService<Prof,Long> {
+public interface ProfService extends BaseService<Prof,Integer> {
 }
